# docs/swagger.yaml
openapi: 3.0.3
info:
  title: FlinkU Project Parking API
  version: 1.0.0
  description: FlinkU 프로젝트 주차장 관련 API 문서

servers:
  - url: http://localhost:5000
    description: Local server

paths:
  /api/parkings:
    get:
      summary: 주차장 목록 조회
      description: 필터, 정렬, 페이지네이션 기능 포함
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: 페이지 번호 (기본값 1)
        - in: query
          name: size
          schema:
            type: integer
          description: 페이지 크기 (기본값 10)
        - in: query
          name: sort
          schema:
            type: string
          description: 정렬 기준 (distance_km 등)
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
          description: 오름차순 / 내림차순
        - in: query
          name: congestion
          schema:
            type: string
            enum: [low, medium, high]
          description: 혼잡도 필터
        - in: query
          name: ev_charger
          schema:
            type: string
            enum: [true, false]
          description: 전기차 충전소 여부 필터
        - in: query
          name: keyword
          schema:
            type: string
          description: 이름 또는 주소 키워드 검색
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: parking 정보를 불러왔습니다.
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: parking
                      total_count:
                        type: integer
                        example: 25
                      page:
                        type: integer
                        example: 1
                      size:
                        type: integer
                        example: 10
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            parking_name:
                              type: string
                              example: 대전역 주차장
                            address:
                              type: string
                              example: 대전 광역시 동구 중앙로 215
                            congestion:
                              type: string
                              example: low
                            ev_charger:
                              type: boolean
                              example: true

  /api/parkings/{id}:
    get:
      summary: 주차장 단일 조회
      description: 주차장 상세 정보 조회
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 조회할 주차장 ID
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 주차장 정보를 불러왔습니다.
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      parking_name:
                        type: string
                        example: 대전역 주차장
                      address:
                        type: string
                        example: 대전광역시 동구 중앙로 214
                      congestion:
                        type: string
                        example: low
                      ev_charger:
                        type: boolean
                        example: true
        '404':
          description: 존재하지 않는 주차장
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: 주차장 정보를 불러올 수 없습니다.
                  error_code:
                    type: string
                    example: NOT_FOUND
  